+++ src/fft.h
@@ -24,15 +24,15 @@
 #define FFT_MAX           2048              // maximum size of FFT transform
 #define WINDOW_MAX        FFT_MAX/2         // maximum size of window
 
-#define OPTIMAL_FFT_SIZE(x)              (2 * pow(2, ceil(log2(x))))
+#define OPTIMAL_FFT_SIZE(x)              ((int) (2 * pow(2, ceil(log2(x)))))
 
-#define WINDOW_SIZE(x,y)                 (floor(((x) * (y) / 1000)))
+#define WINDOW_SIZE(x, y)                 ((size_t) (floor(((x) * (y) / 1000))))
 
 /* Calculate window size and optimal FFT size; FFT size needs to be <= than FFT_MAX (2048 samples)
  * In a case that calculated FFT size is greater than 2048, frame_duration is decreased, until
  * this criteria is met.
  */
-extern int at_calc_window_and_fft_size (int *wsize, int *fft_len, int frame_dur, int srate);
+extern int at_calc_window_and_fft_size(size_t *wsize, int *fft_len, int frame_dur, int srate);
 
 // initialize FFTW library
 extern int at_fftw_init(int size);
@@ -44,28 +44,28 @@ extern int at_fftw_free(void);
 int at_fftw_get_size(void);
 
 // calculate forward FFT transform
-int at_compute_fft(double *time_data_in, int window_size, double *fft_data_out);
+int at_compute_fft(double *time_data_in, size_t window_size, double *fft_data_out);
 
 // calculate inverse FFT transform
-int at_compute_ifft(double *fft_data_in, int window_size, double *time_data_out);
+int at_compute_ifft(double *fft_data_in, size_t window_size, double *time_data_out);
 
 /* calc_magnitude */
-void calc_magnitude (const double * freq, int fft_size, double * magnitude);
+void calc_magnitude(const double *freq, int fft_size, double *magnitude);
 
 /* calc_phase */
-void calc_phase (const double * freq, int fft_size, double * phase);
+void calc_phase(const double *freq, int fft_size, double *phase);
 
 /* calc_power_spectrum */
-double calc_power_spectrum (const double * magnitude, int fft_size, double * power_spectrum);
+double calc_power_spectrum(const double *magnitude, int fft_size, double *power_spectrum);
 
 /* calc_fft_complex_data */
-void calc_fft_complex_data (const double * magnitude, const double * phase, int fft_size, double * freq);
+void calc_fft_complex_data(const double *magnitude, const double *phase, int fft_size, double *freq);
 
 /* multiply_fft_spec_with_gain */
-void multiply_fft_spec_with_gain (const double * gain, int fft_size, double * freq);
+void multiply_fft_spec_with_gain(const double *gain, int fft_size, double *freq);
 
 /* complex argument of FFT spectrum */
-double complex_argument (const double real, const double imag);
+double complex_argument(const double real, const double imag);
 
 
 #endif /* FFT_H_ */
